src/g2r_oscal/requirements.txt
# Using a version of functions-framework that is newer and more stable
functions-framework>=3.5.0
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.49.0
jsonschema